💄 스켈레톤 UI 추가, post-to-card 폴더 추가
