+++ src/pages/post-to/components/post-to-card/SkeletonUI.jsx
@@ -1,0 +1,11 @@
+import SkeletonUIStyle from "./SkeletonUI.style";
+
+const SkeletonUI = () => {
+  return (
+    <SkeletonUIStyle>
+      <div class="loading-img blink_me"></div>
+    </SkeletonUIStyle>
+  );
+};
+
+export default SkeletonUI;
